+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file. This projec
 **Added**
 
 -   Bing support.
+-   Yandex: examples support.
 -   Parsing options.
 
 **Changed**